fix: file recording
Check if file was created before removing it.

+++ Ring/Ring/Features/Conversations/SendFile/SendFileViewModel.swift
@@ -177,6 +177,9 @@ class SendFileViewModel: Stateable, ViewModel {
         }
         self.videoService.videRecordingFinished()
         recordingState.value = .sent
+        if fileName.isEmpty {
+            return
+        }
         try? FileManager.default.removeItem(atPath: fileName)
     }
